+++ examples/resources/meraki_network_client_splash_authorization_status/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_network_client_splash_authorization_status.example "<network_id>,<client_id>"