+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add `meraki_organization_early_access_features` data source
 - Add `meraki_organization_branding_policy` resource and data sources
 - Add `meraki_organization_branding_policies_priorities` resource and data source
+- Add `meraki_network_client_splash_authorization_status` resource and data source
 
 ## 0.1.2